apps/api/requirements.txt
# PhronesisIP — Vercel Serverless Requirements
# Zero cost, API-based stack

# Core
fastapi[all]
uvicorn
gunicorn

# Database
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pgvector  # pgvector support for SQLA

# AI & LLM
groq
langchain-groq
langchain-core
langchain-openai
openai
jina-client  # Jina AI Embeddings v3
httpx
sentence-transformers

# Storage & Auth
boto3        # Cloudflare R2
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings

# Jobs & Ops
inngest
redis        # Upstash Redis for caching
pdfplumber
reportlab    # PDF report generation
sentry-sdk

# Utils
tiktoken
python-multipart
tqdm
